nh-installable: init
