+++ crates/nh-darwin/Cargo.toml
@@ -10,14 +10,15 @@ rust-version.workspace = true
 path = "src/darwin.rs"
 
 [dependencies]
-clap.workspace       = true
-color-eyre.workspace = true
-inquire.workspace    = true
-nh-core.workspace    = true
-nh-remote.workspace  = true
-nix.workspace        = true
-tempfile.workspace   = true
-tracing.workspace    = true
+clap.workspace           = true
+color-eyre.workspace     = true
+inquire.workspace        = true
+nh-core.workspace        = true
+nh-installable.workspace = true
+nh-remote.workspace      = true
+nix.workspace            = true
+tempfile.workspace       = true
+tracing.workspace        = true
 
 [lints]
 workspace = true
