+++ crates/nh-home/Cargo.toml
@@ -10,13 +10,14 @@ rust-version.workspace = true
 path = "src/home.rs"
 
 [dependencies]
-clap.workspace       = true
-color-eyre.workspace = true
-inquire.workspace    = true
-nh-core.workspace    = true
-nh-remote.workspace  = true
-tempfile.workspace   = true
-tracing.workspace    = true
+clap.workspace           = true
+color-eyre.workspace     = true
+inquire.workspace        = true
+nh-core.workspace        = true
+nh-installable.workspace = true
+nh-remote.workspace      = true
+tempfile.workspace       = true
+tracing.workspace        = true
 
 [lints]
 workspace = true
