+++ crates/nh-home/src/home.rs
@@ -9,10 +9,10 @@ use color_eyre::{
 };
 use nh_core::{
   command::{self, Command},
-  installable::{CommandContext, Installable},
   update::update,
   util::{get_hostname, print_dix_diff},
 };
+use nh_installable::{CommandContext, Installable};
 use nh_remote::{self, RemoteBuildConfig};
 use tracing::{debug, info, warn};
