+++ crates/nh-nixos/Cargo.toml
@@ -7,16 +7,17 @@ license.workspace      = true
 rust-version.workspace = true
 
 [dependencies]
-chrono.workspace     = true
-clap.workspace       = true
-color-eyre.workspace = true
-inquire.workspace    = true
-nh-core.workspace    = true
-nh-remote.workspace  = true
-nix.workspace        = true
-serde_json.workspace = true
-tempfile.workspace   = true
-tracing.workspace    = true
+chrono.workspace         = true
+clap.workspace           = true
+color-eyre.workspace     = true
+inquire.workspace        = true
+nh-core.workspace        = true
+nh-installable.workspace = true
+nh-remote.workspace      = true
+nix.workspace            = true
+serde_json.workspace     = true
+tempfile.workspace       = true
+tracing.workspace        = true
 
 [lints]
 workspace = true
