+++ crates/nh-nixos/src/args.rs
@@ -9,8 +9,8 @@ use nh_core::{
     LegacyFeatures,
     OsReplFeatures,
   },
-  installable::Installable,
 };
+use nh_installable::Installable;
 use nh_remote::RemoteHost;
 
 use crate::{
