+++ crates/nh-remote/Cargo.toml
@@ -10,18 +10,19 @@ rust-version.workspace = true
 path = "src/remote.rs"
 
 [dependencies]
-color-eyre.workspace  = true
-indicatif.workspace   = true
-inquire.workspace     = true
-nh-core.workspace     = true
-proptest.workspace    = true
-secrecy.workspace     = true
-serial_test.workspace = true
-shlex.workspace       = true
-signal-hook.workspace = true
-subprocess.workspace  = true
-tracing.workspace     = true
-which.workspace       = true
+color-eyre.workspace     = true
+indicatif.workspace      = true
+inquire.workspace        = true
+nh-core.workspace        = true
+nh-installable.workspace = true
+proptest.workspace       = true
+secrecy.workspace        = true
+serial_test.workspace    = true
+shlex.workspace          = true
+signal-hook.workspace    = true
+subprocess.workspace     = true
+tracing.workspace        = true
+which.workspace          = true
 
 [lints]
 workspace = true
